share button shows action sheet menu. number formatting nicely. deletes are committed.

+++ TapeMeasure/TMNewMeasurementVC.h
@@ -25,7 +25,7 @@
 @interface TMNewMeasurementVC : UIViewController <AVCaptureVideoDataOutputSampleBufferDelegate>
 {
 @private
-	int distanceMeasured;
+	float distanceMeasured;
 	NSTimer *repeatingTimer;
 	bool isMeasuring;
